refactor: service_networking shared utils
